+++ src/app/cart/models/cart-item-interface.ts
@@ -1,6 +1,0 @@
-import { ProductModelInterface } from '../../products/models/product-model-interface';
-
-export interface CartItemInterface {
-  item: ProductModelInterface;
-  count: number;
-}