+++ src/app/cart/models/cart-item.interface.ts
@@ -1,0 +1,6 @@
+import { IProductModel } from '../../products/models/product-model.interface';
+
+export interface ICartItem {
+  item: IProductModel;
+  count: number;
+}