+++ src/app/cart/services/cart.service.ts
@@ -1,17 +1,30 @@
 import { Injectable } from '@angular/core';
-import { CartItemInterface } from '../models/cart-item-interface';
+import { ICartItem } from '../models/cart-item.interface';
+import { IProductModel } from '../../products/models/product-model.interface';
 
 @Injectable({
   providedIn: 'root'
 })
 export class CartService {
 
-  private cartArray: CartItemInterface[] = [];
+  private cartArray: ICartItem[] = [];
 
   constructor() { }
 
-  getProducts(): CartItemInterface[] {
+  getProducts(): ICartItem[] {
     return this.cartArray;
+  }
+
+  addProduct(product: IProductModel): void {
+    const correspondCartItem = this.cartArray.find(x => x.item.id === product.id);
+    if (correspondCartItem) {
+      correspondCartItem.count += 1;
+    } else {
+      this.cartArray.push({
+        count: 1,
+        item: product
+      });
+    }
   }
 
   getCountTotal(): number {